improve package management
