+++ mritopng/models.py
@@ -1,0 +1,6 @@
+class GrayscaleImage(object):
+    
+    def __init__(self, image, width, height):
+        self.image = image
+        self.width = width
+        self.height = height
